pyproject.toml
[project]
name = "semantic-auth"
version = "0.1.0"
description = "Semantic auth pipeline replacing Genie, Knowledge Agent, and Supervisor Agent with direct LLM calls"
requires-python = ">=3.12"
dependencies = [
    "databricks-sdk>=0.74.0",
    "dspy>=3.0.4",
    "neo4j>=5.0.0",
    "pydantic>=2.10.0,<2.12",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
cli = [
    "databricks-job-runner>=0.3.0",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4.0",
]
# MLflow tracing (optional, for --enable-tracing flag)
enrichment = [
    "mlflow[databricks]>=3.7",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/semantic_auth"]
